Moved the handler in it's own file
